change SurveyTemplateController mounting to happen in the survey template module; add researchers to the authorized users of the survey template endpoint; refactor survey-template files to be surveyTemplate
