--- src/survey-template/survey-template.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { SurveyTemplateService } from './survey-template.service';
-
-@Module({
-  providers: [SurveyTemplateService],
-  exports: [SurveyTemplateService],
-})
-export class SurveyTemplateModule {}
